Moved resource costs directly into solace.cmd package.
